Add inline and stack.
