+++ packages/react-ui/src/layout/Inline.ts
@@ -1,0 +1,5 @@
+import { createItemsContainer, ItemsContainerProps } from './internal/createItemsContainer';
+
+export type InlineProps = ItemsContainerProps;
+
+export const Inline = createItemsContainer<InlineProps>('Inline', 'row');
