+++ packages/react-ui/src/layout/Stack.ts
@@ -1,0 +1,5 @@
+import { createItemsContainer, ItemsContainerProps } from './internal/createItemsContainer';
+
+export type StackProps = ItemsContainerProps;
+
+export const Stack = createItemsContainer<StackProps>('Stack', 'column');
